add mha utils
